updated new search design
