+++ app/imports/ui/pages/FindCompanies.jsx
@@ -3,11 +3,10 @@ import { Meteor } from 'meteor/meteor';
 import { Container, Card, Header, Loader } from 'semantic-ui-react';
 import { withTracker } from 'meteor/react-meteor-data';
 import PropTypes from 'prop-types';
-import { Student } from '../../api/student/Student';
-import StudentItem from '../components/StudentItem';
+import { Company } from '../../api/company/Company';
+import CompanyItem from '../components/CompanyItem';
 
-/** Renders a table containing all of the Stuff documents. Use <StuffItem> to render each row. */
-class ListStudents extends React.Component {
+class FindCompanies extends React.Component {
 
   // If the subscription(s) have been received, render the page, otherwise show a loading icon.
   render() {
@@ -17,12 +16,12 @@ class ListStudents extends React.Component {
   // Render the page once subscriptions have been received.
   renderPage() {
     return (
-      <Container id="list-students-page">
-        <Header as="h2" textAlign="center">List Students</Header>
+      <Container id="find-companies-page">
+        <Header as="h2" textAlign="center">Find Companies</Header>
         <Card.Group centered>
-          {this.props.students.map((student, index) => <StudentItem
+          {this.props.companies.map((company, index) => <CompanyItem
             key={index}
-            student={student} />)}
+            company={company} />)}
         </Card.Group>
       </Container>
     );
@@ -30,24 +29,24 @@ class ListStudents extends React.Component {
 }
 
 // Require an array of Stuff documents in the props.
-ListStudents.propTypes = {
-  students: PropTypes.array.isRequired,
+FindCompanies.propTypes = {
+  companies: PropTypes.array.isRequired,
   ready: PropTypes.bool.isRequired,
 };
 
 // withTracker connects Meteor data to React components. https://guide.meteor.com/react.html#using-withTracker
 export default withTracker(() => {
   // Get access to Stuff documents.
-  const subscription = Meteor.subscribe(Student.userPublicationName);
-  const subscription2 = Meteor.subscribe(Student.adminPublicationName);
-  const subscription3 = Meteor.subscribe(Student.recruiterPublicationName);
-  const subscription4 = Meteor.subscribe(Student.studentPublicationName);
+  const subscription = Meteor.subscribe(Company.userPublicationName);
+  const subscription2 = Meteor.subscribe(Company.adminPublicationName);
+  const subscription3 = Meteor.subscribe(Company.studentPublicationName);
+  const subscription4 = Meteor.subscribe(Company.recruiterPublicationName);
   // Determine if the subscription is ready
   const ready = subscription.ready() && subscription2.ready() && subscription3.ready() && subscription4.ready();
   // Get the Stuff documents
-  const students = Student.collection.find({}).fetch();
+  const companies = Company.collection.find({}).fetch();
   return {
-    students,
+    companies,
     ready,
   };
-})(ListStudents);
+})(FindCompanies);
