+++ app/imports/ui/pages/FindStudents.jsx
@@ -3,11 +3,10 @@ import { Meteor } from 'meteor/meteor';
 import { Container, Card, Header, Loader } from 'semantic-ui-react';
 import { withTracker } from 'meteor/react-meteor-data';
 import PropTypes from 'prop-types';
-import { Company } from '../../api/company/Company';
-import CompanyItem from '../components/CompanyItem';
+import { Student } from '../../api/student/Student';
+import StudentItem from '../components/StudentItem';
 
-/** Renders a table containing all of the Stuff documents. Use <StuffItem> to render each row. */
-class ListCompanies extends React.Component {
+class FindStudents extends React.Component {
 
   // If the subscription(s) have been received, render the page, otherwise show a loading icon.
   render() {
@@ -17,12 +16,12 @@ class ListCompanies extends React.Component {
   // Render the page once subscriptions have been received.
   renderPage() {
     return (
-      <Container id="list-companies-page">
-        <Header as="h2" textAlign="center">List Companies</Header>
+      <Container id="find-students-page">
+        <Header as="h2" textAlign="center">Find Students</Header>
         <Card.Group centered>
-          {this.props.companies.map((company, index) => <CompanyItem
+          {this.props.students.map((student, index) => <StudentItem
             key={index}
-            company={company} />)}
+            student={student} />)}
         </Card.Group>
       </Container>
     );
@@ -30,24 +29,24 @@ class ListCompanies extends React.Component {
 }
 
 // Require an array of Stuff documents in the props.
-ListCompanies.propTypes = {
-  companies: PropTypes.array.isRequired,
+FindStudents.propTypes = {
+  students: PropTypes.array.isRequired,
   ready: PropTypes.bool.isRequired,
 };
 
 // withTracker connects Meteor data to React components. https://guide.meteor.com/react.html#using-withTracker
 export default withTracker(() => {
   // Get access to Stuff documents.
-  const subscription = Meteor.subscribe(Company.userPublicationName);
-  const subscription2 = Meteor.subscribe(Company.adminPublicationName);
-  const subscription3 = Meteor.subscribe(Company.studentPublicationName);
-  const subscription4 = Meteor.subscribe(Company.recruiterPublicationName);
+  const subscription = Meteor.subscribe(Student.userPublicationName);
+  const subscription2 = Meteor.subscribe(Student.adminPublicationName);
+  const subscription3 = Meteor.subscribe(Student.recruiterPublicationName);
+  const subscription4 = Meteor.subscribe(Student.studentPublicationName);
   // Determine if the subscription is ready
   const ready = subscription.ready() && subscription2.ready() && subscription3.ready() && subscription4.ready();
   // Get the Stuff documents
-  const companies = Company.collection.find({}).fetch();
+  const students = Student.collection.find({}).fetch();
   return {
-    companies,
+    students,
     ready,
   };
-})(ListCompanies);
+})(FindStudents);
